refactor: router and stories
