+++ src/app/components/DevMenu/index.jsx
@@ -53,7 +53,7 @@ class DevMenu extends React.Component {
 
   render() {
     return (
-      <div className="flex items-center p-3 space-x-3 bg-gray-800 text-white">
+      <div className="flex items-center px-3 py-2 space-x-3 bg-gray-800 text-white">
         <span>Dev Menu</span>
         <button
           className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-1 px-2 rounded"
