--- src/app/components/Welcome/Welcome.jsx
@@ -1,13 +1,0 @@
-import React from "react";
-
-import DevMenu from "../DevMenu";
-
-function Welcome() {
-  return (
-    <div>
-      <DevMenu />
-    </div>
-  );
-}
-
-export default Welcome;
