+++ src/app/router/Popup/styles.scss
