+++ src/app/router/Welcome/index.jsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import ReactDOM from "react-dom";
-import '../../styles/index.css';
+import "../../styles/index.css";
 import Welcome from "./Welcome";
 
 ReactDOM.render(<Welcome />, document.getElementById("welcome-root"));
