--- src/stories/ConnectLnd.stories.js
@@ -1,9 +1,0 @@
-import React from "react";
-import ConnectLnd from "../app/components/Onboard/ConnectLND";
-
-export const Primary = () => <ConnectLnd />;
-
-export default {
-  title: "Pages/Onboard/ConnectLnd",
-  component: ConnectLnd,
-};
