--- src/stories/Laststep.stories.js
@@ -1,9 +1,0 @@
-import React from "react";
-import LastStep from "../app/components/Onboard/LastStep/index";
-
-export const Primary = () => <LastStep />;
-
-export default {
-  title: "Pages/Onboard/LastStep",
-  component: LastStep,
-};
