--- src/stories/Setpassword.stories.js
@@ -1,9 +1,0 @@
-import React from "react";
-import SetPassword from "../app/components/Onboard/SetPassword/index";
-
-export const Primary = () => <SetPassword />;
-
-export default {
-  title: "Pages/Onboard/SetPassword",
-  component: SetPassword,
-};
