--- src/stories/TestConnection.stories.js
@@ -1,9 +1,0 @@
-import React from "react";
-import TestConnection from "../app/components/Onboard/TestConnection";
-
-export const Primary = () => <TestConnection />;
-
-export default {
-  title: "Pages/Onboard/TestConnection",
-  component: TestConnection,
-};
