--- src/stories/Welcome.stories.js
@@ -1,9 +1,0 @@
-import React from "react";
-import WelcomePage from "../app/components/Onboard/Welcome";
-
-export const Primary = () => <WelcomePage />;
-
-export default {
-  title: "Pages/Onboard/Welcome",
-  component: WelcomePage,
-};
