+++ webpack.config.js
@@ -43,10 +43,10 @@ module.exports = {
     background: "./src/extension/background-script/index.js",
     contentScript: "./src/extension/content-script/index.js",
     inpageScript: "./src/extension/inpage-script/index.js",
-    popup: "./src/app/components/Popup/index.jsx",
-    prompt: "./src/app/components/Prompt/index.jsx",
-    options: "./src/app/components/Options/index.jsx",
-    welcome: "./src/app/components/Welcome/index.jsx",
+    popup: "./src/app/router/Popup/index.jsx",
+    prompt: "./src/app/router/Prompt/index.jsx",
+    options: "./src/app/router/Options/index.jsx",
+    welcome: "./src/app/router/Welcome/index.jsx",
     lsat: "./src/extension/ln/lsat/index.js",
   },
